FE: Final Search Implementation
